🧑‍💻(app) add example app
Add an example django app that use the django-peertube-runner-connector
app. This is app should (with peertube runners) be able to do the
entire process of uploading a video, transcoding it, and getting the
result in a directory. This app also has configuration example and
several storages that can be used.