♻️(app) remove request from transcode function
The goal is to simplify the way we are building our urls
that will be use by peertube runners to download a video.
First we remove the need to give a "Request" object to our
transcode function, the current domain is enough.
Second, we remove the job uuid from the url: due to this uuid
being generated at the very last moment, we used a function that
we passed through the whole process that took this job uuid and
created the corresponding url. By removing this, we can now create
the video url at the very start of the process.