✨(app) destination and base_name parameters
Instead of generating files in the same directory of video,
we are adding a parameters permitting to choose where the files will
be created.
Moreover, we also add the possibility to choose the base name of
generated files. The goal is to offer the possibility to not use random
ids that are not practical to use.